tests/wpt/tests/custom-elements/revamped-scoped-registry/Document-createElement.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta name="author" title="Ryosuke Niwa" href="mailto:rniwa@webkit.org">
<link rel="help" href="https://github.com/whatwg/html/issues/10854">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script>

const scopedRegistry = new CustomElementRegistry();
const otherScopedRegistry = new CustomElementRegistry();
class GlobalABElement extends HTMLElement {};
class ScopedABElement extends HTMLElement {};
customElements.define('a-b', GlobalABElement);
scopedRegistry.define('a-b', ScopedABElement);

test(() => {
    assert_true(document.createElement('a-b') instanceof GlobalABElement);
}, 'createElement should use the global registry by default');

test(() => {
    assert_true(document.createElement('a-b', {customElements: scopedRegistry}) instanceof ScopedABElement);
}, 'createElement should use the specified scoped registry');

test(() => {
    const elements = {
        div: HTMLDivElement,
        form: HTMLFormElement,
        span: HTMLSpanElement,
        table: HTMLTableElement,
    };
    for (const localName in elements)
        assert_true(document.createElement(localName, {customElements: scopedRegistry}) instanceof elements[localName], localName);
    for (const localName in elements)
        assert_true(document.createElement(localName, {customElements: window.customElements}) instanceof elements[localName], localName);
}, 'createElement should create a builtin element regardles of a custom element registry specified');

test(() => {
    assert_true(document.createElement('a-b', {customElements: window.customElements}) instanceof GlobalABElement);
}, 'createElement should use the specified global registry');

test(() => {
    const element = document.createElement('a-b', {customElements: otherScopedRegistry});
    assert_equals(element.__proto__.constructor.name, 'HTMLElement');
}, 'createElement should create an upgrade candidate when there is no matching definition in the specified registry');

test(() => {
    class CDElement extends HTMLElement { }
    const registry = new CustomElementRegistry;
    const cdElement = document.createElement('c-d', {customElements: registry});
    assert_false(cdElement instanceof CDElement);
    assert_equals(cdElement.__proto__.constructor.name, 'HTMLElement');
    registry.define('c-d', CDElement);
    assert_false(cdElement instanceof CDElement); // Not yet upgraded since it's disconnected.
    assert_equals(cdElement.__proto__.constructor.name, 'HTMLElement');
    document.body.appendChild(cdElement);
    assert_true(cdElement instanceof CDElement);
}, 'createElement should create an upgrade candidate and the candidate should be upgraded when the element is defined');

</script>
</body>
</html>
