third_party/blink/web_tests/external/wpt/custom-elements/revamped-scoped-registry/Document-createElement.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] createElement should use the specified scoped registry
  assert_true: expected true got false
[FAIL] createElement should create an upgrade candidate when there is no matching definition in the specified registry
  assert_equals: expected "HTMLElement" but got "GlobalABElement"
[FAIL] createElement should create an upgrade candidate and the candidate should be upgraded when the element is defined
  assert_true: expected true got false
Harness: the test ran to completion.
